refactor: move log_config to binary module
